Use the Strategy Pattern in design of this package
Use this pattern to define a family of interchangable algorithms from
which the client code can choose at runtime without having to know
the details of each algorithm's implementation.

Create an interface that declares a method signature that all methods
should implement. Create strategies (modules) for converting each
metadata dialect to implement unique instances of the methods defined in
the interface. Create client code to handle the main workflow logic.