feat: read SSSOM for flow control logic
Read a SSSOM file of a specified metadata standard into a dataframe so
the flow control logic of the convert function may access the match
semantics (predicate) when deciding if a SOSO property meets an
accuracy threshold specified by the user. For example, the convert
function could have a parameter allowing the user to set a match
threshold below which SOSO properties are not created and included in
the final SOSO record.

This allows users to control the accuracy of derived SOSO records to
meet the requirements of specific use cases.