feat: get url property from argument (#5)
Get the SOSO "url" property for the EML strategy from a user supplied
argument, because there is no mapping of this property to the EML
schema, and therefore no way to derive this information from an
EML metadata file.

Use the flexible kwargs parameter as a solution to other unmappable
properties, for the EML strategy and other metadata strategies as well.
Alert users to unmappable properties and their expected value types,
so they may be supplied correctly.

Create a function to test that get_url returns a URL formatted string.