feat: get encodingFormat from file extension (#5)
Retrieve the 'encodingFormat' of a data object based on the MIME type
inferred from the object's file extension. Use this approach since the
'encodingFormat' property does not map to EML.

Please note that this guessing approach may introduce a security concern
when the object's actual encoding format deviates from the guessed
format.