test: strategy methods return None, not null values
Ensure that strategy methods return a None value instead of an null
(empty) value or container. None values are interpreted in the final
graph cleaning step of main.convert as properties with no meaningful
information, which should be removed from the resulting SOSO graph
before being returned to the user.