test: reduce false positive test results
Implement measures to reduce false positive test results in the strategy
interface by explicitly declaring which tests should be skipped, along
with associated rationales. This approach avoids relying on implicit
assumptions represented by a None value.

Originally, a None value was assumed to indicate that a strategy did not
have a method for the property being tested. However, it became
apparent that None could arise in other cases:

1. Method Not Yet Implemented: Used when a strategy method has not
been implemented yet but is planned to be. This tag is applied during
active development of a strategy and should be removed incrementally
as methods are implemented.

2. Property Not in Schema: Applied when the source metadata dialect does
not include content within the schema for the target property the strategy
method is intended to extract. In such cases, the corresponding test is
skipped indefinitely until the metadata dialect schema is updated.

3. Content Missing from Test Metadata: Indicates that although the
property is supported by the metadata's schema, the test metadata is
missing the content and should be added. This should be a temporary
condition.

Implement these measures using the "@pytest.mark.skipif" decorator,
enhancing test reliability and maintainability.