refactor: use method override for unmapped properties
Remove override of SOSO properties via kwargs to guide users to the
more effective and consistent approach of strategy method redefinition
and overrides.

The issue is that some properties cannot be derived from the metadata
file directly but could be supplied through external data. In such
cases, the user would need to pass, and possibly parse additional
information to be included in the SOSO record.

The original thinking was to do this by overriding properties by
defining them in a script and then supplying the values to main.convert
via kwargs. A kwarg name/key matching the top most level property of
the SOSO record (e.g., schema:name) would be replaced by the
corresponding kwarg value.

However, it became apparent that users could do this same task in a way
that utilizes the codebase design, thereby keeping from creating a
separate pathway and unnecessary code for performing this task. This
approach is to create new strategy methods for particular properties and
override the old ones. This approach also encourages users to leverage
the existing implementations in their customizations, which provide
benefit in case of partially mapped properties.

Don't remove kwargs from the strategy interface, only from the kwargs
reference in strategy methods. The strategy interface kwargs is the
pathway by which users pass external data to their custom methods for
processing.

It will be important to create an example of how to do this in the user
documentation.