feat: add get_subject_of method for EML strategy
Add a method to the EML strategy to populate the subjectOf property,
even though some content does not map to EML, thereby leading to a
partially populated property. Do this so clients can leverage this
method when supplying the remaining content.

Note that some of these properties are composed of somewhat arbitrary
values that we provide here for user convenience.