fix: convert SPDX license URL to URI for EML
Convert the schema:license URL to the associated linked data URI for
machine understanding. Implement this as an algorithm that drops the
'.html' suffix of a URL if it belongs to the SPDX vocabulary
(https://spdx.org/licenses/), otherwise return the URL as is.

Note, the SPDX vocabulary is recommended by the EML schema
documentation.