feat: get checksum from EML
Get the checksum method and value from the EML and assign it to the
spdx:checksum property for linked data purposes.

Parsing will be successful only if the "method" attribute of the
authentication element in EML indicates the URI as an SPDX
algorithm. If a non-SPDX algorithm method is listed, return None. Note
that there is no Schema.org property of the DataDownload type that
supports listing an arbitrary authentication method value as a simple
schema:Text (e.g., "MD5").