refactor: remove unused SSSOM functions
Remove unused SSSOM functions from the code base. The original
intent of these functions was to be part of a system that could enable
users to control the accuracy of property method return values based
on the mapping predicate stated in the SSSOM file. This is no longer
within the project scope.