docs: remove "precision-based property control"
Remove references to "precision-based property control" as it is not
yet implemented.

Originally, the plan was to allow users to control the precision of
results from `main.convert` through a "match type threshold"
parameter, which would reference the SSSOM to determine if a
property mapping met the criteria to be included in the returned
graph.