fix: add missing contributors
Add content from EML `contact` and `personnel` elements to the
`contributor` property, as these are currently missing. Exclude
personnel elements of EML `relatedProject` as they are considered
out of scope.

Note: Following SOSO recommendations for linking a `creator` or
`contributor` to their alternate roles requires a PID. Since such PIDs
are not required by the EML schema, this cannot be reliably
implemented.