feat: add spatial coverage support for SPASE (#213)
This is a proposed implementation of `get_spatial_coverage()` strategy
for SPASE, as a list of `schema:Place` objects.

Each `schema:Place` object contains an `identifier` (a URI) and an
`alternateName`, based on the `ObservedRegion` elements present in the
SPASE record. See e.g.:

```
{
  "@type": "schema:Place", 
  "identifier": "http://www.spase-group.org/data/schema/EARTH_MAGNETOSPHERE",
  "alternateName": "Earth.Magnetosphere"
}
```

The `identifier` URI is take from the
[spase.owl](https://github.com/polyneme/topst-spase-rdf-tools/blob/main/data/spase.owl)
draft ontology developed by @polyneme. These URI may be changed in the
future when a formal / open ontology is published for heliophysics.

Note that the ObservedRegions URI are not currently resolvable.