draft of json-ld with edits made by hand
Showing edits here to compare code output with. Should generate this file with code. Note that I pulled the RORs for the University of Iowa and the Southwest Research Institute as examples of RORs in the contributor and creator fields. Please let me know how we can do this for publisher and I will add it to the sssom file.